web/backend/modules/requirements.txt
Flask==0.12.2
requests==2.18.1
tensorflow
keras
numpy
pandas
matplotlib
Pillow==2.2.1
opencv-python
scikit-image
flask-cors
